Refactor frame rotation logic into a dedicated class
Extracted frame rotation functionality into the `AvVideoFrameRotator` class to improve modularity and maintainability. Updated YOLO-related code to utilize the new class and relocated related unit tests to align with the new code structure. This simplifies rotation handling and removes redundant code.
